refactor(api): assemble routes by prefix
X-Anneal-Run: cmthd1kci0x4zmpxnftdaps9s
X-Anneal-Step: 1: Implementation
X-Anneal-Provider: codex
